+++ working/polilinea.h
@@ -8,8 +8,9 @@
 #include <stdbool.h>
 #include <stdint.h>
 #include <stddef.h>
+#include <SDL2/SDL.h>
 #include "color.h"
-
+#include "config.h"
 struct polilinea;
 typedef struct polilinea polilinea_t;
 
@@ -26,7 +27,7 @@ bool polilinea_setear_punto(polilinea_t *polilinea, size_t pos, float x, float y
 bool polilinea_setear_color(polilinea_t *polilinea, color_t color);
 //void polilinea_trasladar(polilinea_t *polilinea, double traslacion[2]);
 //void polilinea_rotar(polilinea_t *polilinea, double angulo);
-bool dibujar_polilinea(polilinea_t *polilinea);
+bool dibujar_polilinea(SDL_Renderer *renderer,polilinea_t *polilinea, float posicion[2], float escala);
 
 //Memoria
 polilinea_t *polilinea_clonar(const polilinea_t *polilinea);
